requirements.txt
yt-dlp
openai-whisper
librosa
numpy
pandas
pytube
ffmpeg
